fix of comments, add onStepChange and write README.md
