+++ components/bl-stepper-component/src/hooks/use-active-class-name.js
@@ -1,0 +1,11 @@
+export const useActiveClassName = (currentStep, stepIndex, completedSteps, stepperClassName) => {
+  let svgIconClass = stepperClassName.titleIcon;
+  let titleTextClass = stepperClassName.titleText;
+
+  if (stepIndex === currentStep || completedSteps.includes(stepIndex)) {
+    svgIconClass += ' ' + stepperClassName.titleIconActive;
+    titleTextClass += ' ' + stepperClassName.titleTextActive;
+  }
+
+  return { svgIconClass, titleTextClass };
+};
