--- components/bl-stepper-component/src/hooks/useStepperClassName.js
@@ -1,44 +1,0 @@
-export const useStepperClassName = (typeStepper) => {
-  switch (typeStepper) {
-    case 'root':
-      return {
-        stepper: 'stepper',
-        item: 'stepper__item',
-        stepTitle: 'stepper__step-title step-title',
-        line: 'stepper__line',
-        titleIcon: 'step-title__icon',
-        titleIconActive: 'step-title__icon--active',
-        titleIconText: 'step-title__icon-text',
-        titleText: 'step-title__text',
-        titleTextActive: 'step-title__text--active',
-      }
-
-    case 'alternativeLabel':
-      return {
-        stepper: 'stepper-alternative-label',
-        item: 'stepper-alternative-label__item',
-        stepTitle: 'stepper-alternative-label__step-title step-alternative-label-title',
-        line: 'stepper-alternative-label__line',
-        titleIcon: 'step-title__icon',
-        titleIconActive: 'step-title__icon--active',
-        titleIconText: 'step-title__icon-text',
-        titleText: 'step-alternative-label-title__text',
-        titleTextActive: 'step-alternative-label-title__text--active',
-      }
-
-    case 'customized': {
-      return {
-        stepper: 'stepper',
-        item: 'stepper-alternative-label__item',
-        stepTitle: 'stepper-alternative-label__step-title step-alternative-label-title',
-        line: 'stepper-customized__line',
-        lineActive: 'stepper-customized__line--active',
-        titleIcon: 'step-customized-title__icon',
-        titleIconActive: 'step-customized-title__icon--active',
-        titleIconComplete: 'step-customized-title__icon-complete',
-        titleText: 'step-alternative-label-title__text',
-        titleTextActive: 'step-alternative-label-title__text--active',
-      }
-    }
-  }
-}
