+++ components/bl-stepper-component/styles/index.less
@@ -18,173 +18,176 @@
 @bl-customComponent-stepper-customized-iconSize: 18px;
 @bl-customComponent-stepper-customized-iconMarkSize: 8px;
 
-.stepper {
-  display: flex;
-  align-items: center;
-
-  &__item {
+.bl-customComponent-stepper {
+  .stepper {
     display: flex;
     align-items: center;
 
-    &:not(:last-child) {
-      width: @bl-customComponent-stepper-itemWidth;
-    }
-  }
-
-  &__step-title {
-    margin-right: 7px;
-  }
-
-  &__line {
-    display: flex;
-    flex: 1 1 auto;
-
-    margin-right: 7px;
-
-    border-color: @bl-customComponent-stepper-lineColor;
-    border-top-style: solid;
-    border-top-width: @bl-customComponent-stepper-lineBorderWidth;
-  }
-}
-
-.step-title {
-  display: flex;
-  align-items: center;
-
-  &__icon {
-    display: inline-flex;
-    justify-content: center;
-    align-items: center;
-
-    width: @bl-customComponent-stepper-iconSize;
-    height: @bl-customComponent-stepper-iconSize;
-
-    border-radius: @bl-customComponent-stepper-iconBorderRadius;
-    fill: @bl-customComponent-stepper-textColor;
-    opacity: 0.38;
-
-    transition: fill @bl-customComponent-stepper-transition;
-
-    &--active {
-      fill: @bl-customComponent-stepper-themeColor;
-      opacity: 1;
-    }
-
-    &-text {
-      fill: @bl-customComponent-stepper-backgroundColor;
-      transition: fill @bl-customComponent-stepper-transition;
-    }
-  }
-
-  &__text {
-    display: block;
-    opacity: 0.6;
-
-    margin-left: 5px;
-
-    transition: opacity @bl-customComponent-stepper-transition;
-
-    &--active {
-      opacity: 1;
-    }
-  }
-}
-
-.stepper-alternative-label {
-  display: flex;
-
-  &__item {
-    position: relative;
-
-    display: flex;
-    flex: 1;
-    justify-content: center;
-  }
-
-  &__line {
-    display: block;
-
-    position: absolute;
-    top: 12px;
-    left: calc(50% + 20px);
-    right: calc(-50% + 20px);
-
-    border-color: @bl-customComponent-stepper-lineColor;
-    border-top-style: solid;
-    border-top-width: @bl-customComponent-stepper-lineBorderWidth;
-  }
-}
-
-.step-alternative-label-title {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-
-  &__text {
-    display: block;
-    opacity: 0.6;
-
-    margin-top: @bl-customComponent-stepper-alternative-label-titleTextMarginTop;
-
-    transition: opacity @bl-customComponent-stepper-transition;
-
-    &--active {
-      opacity: 1;
-    }
-  }
-}
-
-.step-customized-title {
-  &__icon {
-    width: @bl-customComponent-stepper-customized-iconMarkSize;
-    height: @bl-customComponent-stepper-customized-iconMarkSize;
-    border-radius: 50%;
-    background-color: @bl-customComponent-stepper-textColor;
-    opacity: 0.38;
-    transition: background-color @bl-customComponent-stepper-transition;
-
-    &--active {
-      background-color: @bl-customComponent-stepper-themeColor;
-      opacity: 1;
-    }
-
-    &-mark-container {
+    &__item {
       display: flex;
+      align-items: center;
 
+      &:not(:last-child) {
+        width: @bl-customComponent-stepper-itemWidth;
+      }
+    }
+
+    &__step-title {
+      margin-right: 7px;
+    }
+
+    &__line {
+      display: flex;
+      flex: 1 1 auto;
+
+      margin-right: 7px;
+
+      border-color: @bl-customComponent-stepper-lineColor;
+      border-top-style: solid;
+      border-top-width: @bl-customComponent-stepper-lineBorderWidth;
+    }
+  }
+
+  .step-title {
+    display: flex;
+    align-items: center;
+
+    &__icon {
+      display: inline-flex;
       justify-content: center;
       align-items: center;
 
-      height: @bl-customComponent-stepper-customized-iconSize;
-      width: @bl-customComponent-stepper-customized-iconSize;
-    }
+      width: @bl-customComponent-stepper-iconSize;
+      height: @bl-customComponent-stepper-iconSize;
 
-    &-complete {
-      height: @bl-customComponent-stepper-customized-iconSize;
-      width: @bl-customComponent-stepper-customized-iconSize;
+      border-radius: @bl-customComponent-stepper-iconBorderRadius;
+      fill: @bl-customComponent-stepper-textColor;
+      opacity: 0.38;
 
-      fill: @bl-customComponent-stepper-themeColor;
       transition: fill @bl-customComponent-stepper-transition;
+
+      &--active {
+        fill: @bl-customComponent-stepper-themeColor;
+        opacity: 1;
+      }
+
+      &-text {
+        fill: @bl-customComponent-stepper-backgroundColor;
+        transition: fill @bl-customComponent-stepper-transition;
+      }
+    }
+
+    &__text {
+      display: block;
+      opacity: 0.6;
+
+      margin-left: 5px;
+
+      transition: opacity @bl-customComponent-stepper-transition;
+
+      &--active {
+        opacity: 1;
+      }
+    }
+  }
+
+  .stepper-alternative-label {
+    display: flex;
+
+    &__item {
+      position: relative;
+
+      display: flex;
+      flex: 1;
+      justify-content: center;
+    }
+
+    &__line {
+      display: block;
+
+      position: absolute;
+      top: 12px;
+      left: calc(50% + 20px);
+      right: calc(-50% + 20px);
+
+      border-color: @bl-customComponent-stepper-lineColor;
+      border-top-style: solid;
+      border-top-width: @bl-customComponent-stepper-lineBorderWidth;
+    }
+  }
+
+  .step-alternative-label-title {
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+
+    &__text {
+      display: block;
+      opacity: 0.6;
+
+      margin-top: @bl-customComponent-stepper-alternative-label-titleTextMarginTop;
+
+      transition: opacity @bl-customComponent-stepper-transition;
+
+      &--active {
+        opacity: 1;
+      }
+    }
+  }
+
+  .step-customized-title {
+    &__icon {
+      width: @bl-customComponent-stepper-customized-iconMarkSize;
+      height: @bl-customComponent-stepper-customized-iconMarkSize;
+      border-radius: 50%;
+      background-color: @bl-customComponent-stepper-textColor;
+      opacity: 0.38;
+      transition: background-color @bl-customComponent-stepper-transition;
+
+      &--active {
+        background-color: @bl-customComponent-stepper-themeColor;
+        opacity: 1;
+      }
+
+      &-mark-container {
+        display: flex;
+
+        justify-content: center;
+        align-items: center;
+
+        height: @bl-customComponent-stepper-customized-iconSize;
+        width: @bl-customComponent-stepper-customized-iconSize;
+      }
+
+      &-complete {
+        height: @bl-customComponent-stepper-customized-iconSize;
+        width: @bl-customComponent-stepper-customized-iconSize;
+
+        fill: @bl-customComponent-stepper-themeColor;
+        transition: fill @bl-customComponent-stepper-transition;
+      }
+    }
+  }
+
+  .stepper-customized__line {
+    display: block;
+
+    opacity: 0.38;
+    border-color: @bl-customComponent-stepper-textColor;
+    border-top-style: solid;
+    border-top-width: @bl-customComponent-stepper-customized-lineBorderWidth;
+
+    position: absolute;
+    top: 8px;
+    left: calc(50% + 20px);
+    right: calc(-50% + 20px);
+
+    transition: border-color @bl-customComponent-stepper-transition;
+
+    &--active {
+      border-color: @bl-customComponent-stepper-themeColor;
+      opacity: 1;
     }
   }
 }
 
-.stepper-customized__line {
-  display: block;
-
-  opacity: 0.38;
-  border-color: @bl-customComponent-stepper-textColor;
-  border-top-style: solid;
-  border-top-width: @bl-customComponent-stepper-customized-lineBorderWidth;
-
-  position: absolute;
-  top: 8px;
-  left: calc(50% + 20px);
-  right: calc(-50% + 20px);
-
-  transition: border-color @bl-customComponent-stepper-transition;
-
-  &--active {
-    border-color: @bl-customComponent-stepper-themeColor;
-    opacity: 1;
-  }
-}
